+++ example/dynamics/main.tf
@@ -1,0 +1,18 @@
+resource "aws_s3_bucket" "this" {
+
+
+  dynamic "website" {
+    for_each = length(keys(var.website)) == 0 ? [] : [var.website]
+
+    content {
+      index_document           = lookup(website.value, "index_document", null)
+      error_document           = lookup(website.value, "error_document", null)
+      redirect_all_requests_to = lookup(website.value, "redirect_all_requests_to", null)
+      routing_rules            = lookup(website.value, "routing_rules", null)
+    }
+  }
+}
+
+resource "aws_s3_bucket" "that"{
+    name = concat(aws_s3_bucket.website.index_document,"")
+}
