+++ internal/app/tfsec/rules/aws/apigateway/use_secure_tls_policy_rule.go
@@ -55,6 +55,7 @@ resource "aws_api_gateway_domain_name" "good_example" {
 					WithDescription("Resource '%s' should include security_policy (defaults to outdated SSL/TLS policy).", resourceBlock.FullName())
 				return
 			}
+
 			if securityPolicyAttr.NotEqual("TLS_1_2") {
 				set.AddResult().
 					WithDescription("Resource '%s' defines outdated SSL/TLS policies (not using TLS_1_2).", resourceBlock.FullName()).
